JS file login_logout.js has been implemented
