+++ web/src/main/java/com/isa/zajavieni/rest/ApiController.java
@@ -5,5 +5,4 @@ import javax.ws.rs.core.Application;
 
 @ApplicationPath("/api")
 public class ApiController extends Application {
-
 }
