+++ web/src/main/java/com/isa/zajavieni/servlet/MainServlet.java
@@ -34,6 +34,7 @@ public class MainServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
         List<EventDto> events = eventDtoService.findUpcomingEvents(0, EVENETS_ON_THE_WELCOME);
         Template template = templateProvider.getTemplate(getServletContext(), "welcome-page.ftlh");
         Map<String, Object> model = new HashMap<>();
