--- web/src/main/java/com/isa/zajavieni/servlet/UserManagerServlet.java
@@ -1,6 +1,0 @@
-package com.isa.zajavieni.servlet;
-
-
-public class UserManagerServlet {
-
-}
