+++ web/src/main/webapp/WEB-INF/fm-templates/navbar.ftlh
@@ -41,8 +41,10 @@
             </#if>
             <a class="p-2 text-dark" href="#">Koszyk</a>
 
-            <a href="/login" class="btn btn-danger btn-sm"><i class="fa fa-google"></i> Zaloguj się przez <b>Google</b></a>
-            <a href="/logout" class="btn btn-danger btn-sm"><i class="fa fa-google"></i> Wyloguj się</a>
+            <a href="/login" class="btn btn-danger btn-sm" id="login">
+                <i class="fa fa-google"></i> Zaloguj się przez <b>Google</b></a>
+            <a href="/logout" class="btn btn-danger btn-sm"
+               id="logout"><i class="fa fa-google"></i> Wyloguj się</a>
             <#--  <a class="p-2 text-dark" href="/admin/upload">Admin</a>-->
         </nav>
     </div>
@@ -58,6 +60,7 @@
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
 <script src="https://ajax.googleapis.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min.js"></script>
 <script src="/js/permissions.js"></script>
+<script src="/js/login_logout.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
         integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
         crossorigin="anonymous"></script>
